Implement equality testing for classes with _ptr pointers
This implements __eq__ in a Ptr base class that is passed to other
classes that can use it to test between their respective `self._ptr`
attributes. This means that new class instances created as part of class
properties (e.g. `Seat.keyboard` which creates a new instance of
`Keyboard`) can be compared with existing instances that represent the
same thing (i.e. have the same pointer).