Add an interface for wlr_output_management_v1.h
This adds handlers for the various structs and methods in wlroots'
wlr_output_management_v1.h. This lets compositors create output managers
and process client requests to change the output layout and
configuration of each output. The output layout and output interfaces
have been extended somewhat to allow for the interaction between these
and the output manager objects.