Add automatic wlr_texture_destroy gc handling
For textures that are allocated via wlr_texture_from_pixels, setup the
ptr to be destroyed via wlr_texture_destroy when it is deleted.  Use
`gc` to ensure this is always called.  Setup the `destroy` method to
invoke this gc method when it is setup, but should be a no-op for
objects that are not allocated in this way.