Use new type syntax
Modify the type annotations to use the new Python 3.10 syntax for types,
in particular using the "X | Y" union types and using the builtin
generic types rather than the `typing` equivalents.  Without adding a
lot of quoted type annotations (which would make this a net negative
change, IMO), this requires dropping Python 3.6 support to be able to
use the `__future__` annotations functionality.