Fix installation problems
Fix up the setup.py file so the cffi package and the required
cffi-related dependencies are specified as setup_requires. To keep from
needing to install these packages in all cases, make the setup dynamic
based on the arguments used to invoke it. This allows running help or
doing `sdist` without needing the dependencies installed. More can be
added here if needed, but this should be sufficient for the core CI
functionality to work.

Fixes #65