Don't wrap IdleInhibitorV1's destroy event data
According to wlroots' emitter for this signal, it *does* return the idle
inhibitor's `struct wlr_surface*`. However, in practice the pointed-to
data seems to be invalid, causing a crash (e.g.
https://github.com/qtile/qtile/issues/3361).

The data does not seem to be used by any compositor so until it is
needed we can probably get away without the wrapper, resolving the
problem.