tinywl: use scene graph for view_at()
This uses the scene graph api for the compositor's view_at() method,
with logic lifted from the original wlroots' tinywl compositor. This
makes sure that all nodes in the scene are considered when using the
method. As an example of what this fixes, try hoving the pointer over
the client-side title bar of `foot` before and after this commit.